+++ src/ngme.cpp
@@ -1,6 +1,7 @@
 // implement the Ngme class and rand effect class
 #include "ngme.h"
 
+#include "include/thread_io.h"
 #include <atomic>
 
 #ifdef _OPENMP
@@ -195,7 +196,7 @@ void Ngme::set_parameter_and_update(const VectorXd &p, bool with_precond) {
 
   // set the different parameter for each random effect
   if (debug)
-    std::cout << "set_parameter() in ngme class" << std::endl;
+    ngme_io::out() << "set_parameter() in ngme class" << std::endl;
   // Invalidate caches
   grad_valid_ = false;
   precond_valid_ = false;